refact : BlackJackGamer open 클래스로 변경
